fix: Resolve 21 SonarCloud issues and eliminate code duplication
- Fixed assertion argument order in test classes (assertNotEquals)
- Removed redundant assertNotNull after assertThrows
- Added @SuppressWarnings for equals() implementation tests
- Created AbstractStubRepository to eliminate 3.53% code duplication
- Reduced instruction count from 966 to 948 (-18 lines)
- All 794 tests passing, coverage maintained at 98%/97%

SonarCloud issues resolved: 21
Quality gate: Expected to pass (duplication <3.0%)
