+++ noakweather-platform/weather-storage/src/main/java/weather/storage/repository/dynamodb/DynamoDbRepository.java
@@ -16,20 +16,13 @@
  */
 package weather.storage.repository.dynamodb;
 
-import weather.model.WeatherData;
-import weather.model.WeatherDataSource;
-import weather.storage.repository.RepositoryStats;
-import weather.storage.repository.UniversalWeatherRepository;
-
+import weather.storage.repository.AbstractStubRepository;
 import java.time.LocalDateTime;
-import java.util.Collections;
-import java.util.List;
-import java.util.Optional;
 
 /**
  * DynamoDB implementation of the UniversalWeatherRepository.
  * 
- * STUB IMPLEMENTATION - Week 1, Day 5
+ * STUB IMPLEMENTATION
  * This is a placeholder for future DynamoDB integration in the speed/serving layer.
  * 
  * DynamoDB serves as the real-time data store in Lambda Architecture:
@@ -52,98 +45,23 @@ import java.util.Optional;
  * - Add CloudWatch metrics integration
  * 
  * @author bclasky1539
- *
  */
-public class DynamoDbRepository implements UniversalWeatherRepository {
+public class DynamoDbRepository extends AbstractStubRepository {
     
-    private static final String NOT_IMPLEMENTED_MESSAGE = 
-        "DynamoDbRepository not yet implemented - scheduled for later";
-    
+    /**
+     * Creates a new DynamoDbRepository stub.
+     */
     public DynamoDbRepository() {
-        // Stub constructor - no initialization yet
-    }
-    
-    @Override
-    public WeatherData save(WeatherData weatherData) {
-        // Will use: PutItemRequest with partition key (station_id) and sort key (observation_time)
-        // Add TTL attribute for automatic expiration
-       throw new UnsupportedOperationException(NOT_IMPLEMENTED_MESSAGE);
-    }
-    
-    @Override
-    public int saveBatch(List<WeatherData> weatherDataList) {
-        // Will batch in groups of 25 (DynamoDB limit)
-        // Handle unprocessed items with exponential backoff
-        throw new UnsupportedOperationException(NOT_IMPLEMENTED_MESSAGE);
-    }
-    
-    @Override
-    public Optional<WeatherData> findByStationAndTime(String stationId, LocalDateTime observationTime) {
-        // Will use: GetItemRequest with partition key + sort key
-        // This is the most efficient query pattern in DynamoDB
-        return Optional.empty();
-    }
-    
-    @Override
-    public List<WeatherData> findByStationAndTimeRange(String stationId, 
-                                                        LocalDateTime startTime, 
-                                                        LocalDateTime endTime) {
-        // Will use: QueryRequest with partition key and sort key BETWEEN condition
-        // Efficient: Uses the primary key's sort key for range queries
-        return Collections.emptyList();
-    }
-    
-    @Override
-    public Optional<WeatherData> findLatestByStation(String stationId) {
-        // Will use: QueryRequest with partition key, ScanIndexForward=false, Limit=1
-        // Retrieves the most recent item efficiently
-        return Optional.empty();
-    }
-    
-    @Override
-    public List<WeatherData> findBySourceAndTimeRange(WeatherDataSource source,
-                                                       LocalDateTime startTime,
-                                                       LocalDateTime endTime) {
-        // Will use: QueryRequest on GSI with data_source as partition key
-        // Note: Less efficient than primary key queries, but necessary for multi-source support
-        return Collections.emptyList();
-    }
-    
-    @Override
-    public List<WeatherData> findByStationsAndTime(List<String> stationIds, 
-                                                    LocalDateTime observationTime) {
-        // Will use: BatchGetItemRequest with multiple keys (station_id + observation_time)
-        // Batch in groups of 100 (DynamoDB limit)
-        return Collections.emptyList();
+        super("DynamoDbRepository");
     }
     
     @Override
     public int deleteOlderThan(LocalDateTime cutoffDate) {
+        // Override with specific message about TTL
         // TTL is more cost-effective and automatic in DynamoDB
         // Manual deletion would require: Scan + BatchWriteItem (expensive!)
         throw new UnsupportedOperationException(
             "DynamoDbRepository.deleteOlderThan() - use TTL for automatic expiration instead"
-        );
-    }
-    
-    @Override
-    public boolean isHealthy() {
-        // Will test: DynamoDB client connectivity, DescribeTable call
-        return false; // Not operational yet
-    }
-    
-    @Override
-    public RepositoryStats getStats() {
-        // Note: DynamoDB doesn't provide COUNT efficiently - may need to maintain counters
-        // Consider using DynamoDB Streams + Lambda to track statistics
-        return new RepositoryStats(
-            0L,           // totalRecordCount (approximate - DynamoDB doesn't track precisely)
-            null,         // oldestRecordTime
-            null,         // newestRecordTime
-            0L,           // recordsLast24Hours
-            0L,           // recordsLast7Days
-            0,            // uniqueStationCount
-            0L            // storageSize (from DescribeTable)
         );
     }
 }
