+++ noakweather-platform/weather-storage/src/main/java/weather/storage/repository/snowflake/SnowflakeRepository.java
@@ -16,20 +16,12 @@
  */
 package weather.storage.repository.snowflake;
 
-import weather.model.WeatherData;
-import weather.model.WeatherDataSource;
-import weather.storage.repository.RepositoryStats;
-import weather.storage.repository.UniversalWeatherRepository;
-
-import java.time.LocalDateTime;
-import java.util.Collections;
-import java.util.List;
-import java.util.Optional;
+import weather.storage.repository.AbstractStubRepository;
 
 /**
  * Snowflake implementation of the UniversalWeatherRepository.
  * 
- * STUB IMPLEMENTATION - Week 1, Day 5
+ * STUB IMPLEMENTATION
  * This is a placeholder for future Snowflake integration in the batch layer.
  * 
  * Snowflake serves as the immutable master data store in Lambda Architecture:
@@ -44,94 +36,13 @@ import java.util.Optional;
  * - Implement proper error handling and retry logic
  * 
  * @author bclasky1539
- *
  */
-public class SnowflakeRepository implements UniversalWeatherRepository {
+public class SnowflakeRepository extends AbstractStubRepository {
     
-    private static final String NOT_IMPLEMENTED_MESSAGE = 
-        "SnowflakeRepository not yet implemented - scheduled for later";
-    
+    /**
+     * Creates a new SnowflakeRepository stub.
+     */
     public SnowflakeRepository() {
-        // Stub constructor - no initialization yet
-    }
-    
-    @Override
-    public WeatherData save(WeatherData weatherData) {
-        // Will use: INSERT INTO weather_data VALUES (?, ?, ?, ...)
-        throw new UnsupportedOperationException(NOT_IMPLEMENTED_MESSAGE);
-    }
-    
-    @Override
-    public int saveBatch(List<WeatherData> weatherDataList) {
-        // Will use: COPY INTO or batch INSERT for efficiency
-        throw new UnsupportedOperationException(NOT_IMPLEMENTED_MESSAGE);
-    }
-    
-    @Override
-    public Optional<WeatherData> findByStationAndTime(String stationId, LocalDateTime observationTime) {
-        // Will use: SELECT * FROM weather_data WHERE station_id = ? AND observation_time = ?
-        return Optional.empty();
-    }
-    
-    @Override
-    public List<WeatherData> findByStationAndTimeRange(String stationId, 
-                                                        LocalDateTime startTime, 
-                                                        LocalDateTime endTime) {
-        // Will use: SELECT * FROM weather_data 
-        //           WHERE station_id = ? AND observation_time BETWEEN ? AND ?
-        //           ORDER BY observation_time
-        return Collections.emptyList();
-    }
-    
-    @Override
-    public Optional<WeatherData> findLatestByStation(String stationId) {
-        // Will use: SELECT * FROM weather_data 
-        //           WHERE station_id = ? 
-        //           ORDER BY observation_time DESC 
-        //           LIMIT 1
-        return Optional.empty();
-    }
-    
-    @Override
-    public List<WeatherData> findBySourceAndTimeRange(WeatherDataSource source,
-                                                       LocalDateTime startTime,
-                                                       LocalDateTime endTime) {
-        // Will use: SELECT * FROM weather_data 
-        //           WHERE data_source = ? AND observation_time BETWEEN ? AND ?
-        return Collections.emptyList();
-    }
-    
-    @Override
-    public List<WeatherData> findByStationsAndTime(List<String> stationIds, 
-                                                    LocalDateTime observationTime) {
-        // Will use: SELECT * FROM weather_data 
-        //           WHERE station_id IN (?, ?, ...) AND observation_time = ?
-        return Collections.emptyList();
-    }
-    
-    @Override
-    public int deleteOlderThan(LocalDateTime cutoffDate) {
-        // Will use: DELETE FROM weather_data WHERE observation_time < ?
-       throw new UnsupportedOperationException(NOT_IMPLEMENTED_MESSAGE);
-    }
-    
-    @Override
-    public boolean isHealthy() {
-        // Will test: Connection pool status, simple SELECT 1 query
-        return false; // Not operational yet
-    }
-    
-    @Override
-    public RepositoryStats getStats() {
-        // Will query: COUNT(*), MIN/MAX observation_time, COUNT(DISTINCT station_id), etc.
-        return new RepositoryStats(
-            0L,           // totalRecordCount
-            null,         // oldestRecordTime
-            null,         // newestRecordTime
-            0L,           // recordsLast24Hours
-            0L,           // recordsLast7Days
-            0,            // uniqueStationCount
-            0L            // storageSize
-        );
+        super("SnowflakeRepository");
     }
 }
